[frontend] Update client list and apply some minor fixes (#734)
* Add logos for Fedilab, Toot and Mastodon to assets folder

Adds multiple logo's for the changes on the client list.

* Add clients to the index page to give users a wider choice and some minor fixes

Adds Fedilab, Toot! and "More clients" to the client list.
Also adds periods after all descriptions, so they are unified in writing style.
Adds target blank and rel noopener to external links.
Changes the pinafore.social link to link to the homepage, like the other clients.

* Add 'the' to the description text

Adds the missing 'the' to the More clients description

* Delete fedilab.png

* Delete toot.png

* Remove added clients, but keep More clients text.

Removes the Fedilab and Toot! clients.