Create login.md
as many people fail to set up _toot cli_ client, I've this detailed guide about login and clients so any can reference to. It's adapted from my blog post and of course feel free to adapt and/or correct to the community guides and styling.

I can not upload images to `assets` so I linked to my blog

I hope it's useful.