dont bundle assets in releases, rework docker builds
we now need to build the assets before we build the executable.
this means that the current approach of the docker builds is not
working anymore. previously, the dockerfile tried to accomodate for
standalone and goreleaser builds, ending up in duplicate work for
goreleaser and still not providing a fully selfcontained build for
standalone builds.
this commit separates both use cases, now having two dockerfiles:

- prebuilt.Dockerfile is for goreleaser, it just copies in the built
  binary
- Dockerfile now contains the whole build process, is truly
  selfcontained.

i tried to keep it DRY, but that's not really feasible with dockerfiles
without custom syntax extensions.