[feat] Add CodeQL analysis
This is configured to run weekly, at 05:42 UTC on Mondays. It can also
be triggered by hand. It's not enabled on a per-PR case because
extracting all the types from the activity package takes north of 5min,
making a whole analysis run take a litte under 8min. That seems annoying
to potentially wait for on a PR.