[feature] Add support for edited statuses
This adds receive support for the "status.update" type used by Mastodon
and possibly other ActivityPub implementations for allowing
Notes/Statuses to be edited.

This does not add the ability to edit posts on GTS itself, but does
properly update an edited status in the timeline and send the edit
information via webhook connections.