[feature] Make client IP logging configurable
Admins can now chose whether client IPs should be present in the logs.
By default this is enabled in order to retain the current behaviour and
because when running in non-proxied mode this is something admins
probably want.

When having an LB in front of GtS this can be disabled as requsets can
still be correlated between LB and GtS using the Request ID or the
tracing facilities.

Fixes: #1574