[feature] Treat unlisted as public for visibility
Up to now, GtS has required authentication to view an unlisted/unlocked
status. However, looking around other implementations unlisted/unlocked
still means public, but with the caveat that it won't show up in local
and federated timelines or in hashtag searches. It's very often used to
reply to a thread to avoid filling the timelines with replies to a
conversation someone might not be interested in following.

Changing this will now make unlisted replies to a status on this
instance viewable, which means we get a more complete view in the thread
UI in our web UI and should result in a more complete view of a thread
in other client applications too.