[feature] Don't hardcode membership claim
In order to solve #1914 and #1915 we need to make our matching on OIDC
claims more flexible. Unfortunately, there is no well-defined and
supported scope amongst all IdPs to expose membership information using
defined claims. So far we've expected a groups claim to exist, but this
is implementation dependent.

This changes our claims implementation so that for standardised claims
defined by a well-known scope we retain a struct field. Any other keys
in the claims object are stored by their key in the new Attrs field
instead, as a list of strings. A custom UnmarshalJSON is implemented for
this purpose in order to ensure we have type safety once the decoding is
done.