[feature] Allow flexible OIDC claim matching
This introduces 2 new options, oidc-admin-required-claims and
oidc-account-required-claims. This lets an admin define claim/value
pairs that the claims on the ID token will be tested against. In our
implementation, all claims need to match in order to clear the
requirement.

There is a key difference between oidc-account-required-claims and
oidc-admin-required claims, in that the former may be ommitted entirely
and then the test will be skipped. This means no claim check has to be
configured in order to allow for account creation/login. For
oidc-admin-required-claims, admin access is never granted if no claims
are configured.