[feature] Allow exposing lookup without auth
This lets an instance admin configure whether the
/api/v1/accounts/lookup endpoint requires authentication or not. Parts
of this information can be gleamed through webfinger so it's conceivable
people are OK with this being public.

The endpoint is documented as being public in the Masto API, so this
allows an admin to make our implementation behave in the same manner
should they want to.

The default remains false, it needs to be explicitly enabled.

Fixes #1939