[feature] Use Read/Write lock for config
The most common case for GtS is to read config at startup and never call
the config.SetXXX methods after that. This means that in almsot all
cases when GtS is running, the only thing we'll do is call
config.GetXXX. However, since we currently use a Lock(), two
config.GetXXX methods for the same value will have to wait on each
other.

This changes the config state mutex to an RWMutex, allowing many
goroutines to read the config at the same time. This should ensure that
requests in HTTP handlers aren't blocked on eachother if they happen to
call the same config.GetXXX method.

The Set methods continue to call Lock() instead of RLock(), as they need
to modify the config state, and the same goes for Viper() and Config().