[feature] Report Masto version in /api/v1/instance
A number of applications use the version reported by /api/v1/instance as
a way to do feature detection for the Mastodon client API. Due to the
current GtS versions in 0.x, this often means GtS gets determined to be
non-compatible by clients and libraries that only ever test against
Mastodon proper.

This adds a configuration value that allows the admin to decide to
inject a Mastodon version in that response instead, in order to cajool
clients into working correctly with GtS. This is only done for
/api/v1/instance as that's the Masto client API, nodeinfo is left as it
is as that should not be used by Masto clients for feature detection.

Since the clients we recommend don't need to be tricked this way, this
feature defaults to being off. Admins can enable it if so desired, which
may be useful for small communities where admins might have less say
over which clients people use.