[feature] Allow admin to always create accounts
This lets an admin create accounts using the client API, even if server
registrations are globally disabled. When registration is disabled, the
presented token is checked for being a user and if the user is an admin,
the registration is still allowed to proceed.

This matches the behaviour on the CLI, as even when registrations are
closed an admin can currently still create new accounts.